refactor: component binding exists in common
